yay its all starting to come together now
