+++ package.json
@@ -56,7 +56,6 @@
     "eslint-plugin-jest": "^22.15.2",
     "eslint-plugin-react": "^7.14.3",
     "file-loader": "^4.2.0",
-    "google-fonts-webpack-plugin": "^0.4.4",
     "html-webpack-plugin": "^3.2.0",
     "husky": "^3.0.4",
     "image-webpack-loader": "^5.0.0",
