--- ui/package.json
@@ -1,60 +1,0 @@
-{
-  "name": "ui",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "jest",
-    "dev": "npx webpack-dev-server --config webpack.dev.js",
-    "lint": "eslint .",
-    "lint:fix": "eslint . --fix",
-    "build": "webpack --config webpack.common.js"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@babel/polyfill": "^7.4.0",
-    "babel-eslint": "^10.0.2",
-    "bootstrap": "^4.2.1",
-    "eslint": "^6.1.0",
-    "eslint-plugin-jest": "^22.14.1",
-    "eslint-plugin-react": "^7.14.3",
-    "i18next": "^17.0.6",
-    "i18next-browser-languagedetector": "^3.0.1",
-    "node-sass-glob-importer": "^5.3.1",
-    "query-string": "^6.8.1",
-    "react": "^16.6.3",
-    "react-bootstrap": "^1.0.0-beta.6",
-    "react-dom": "^16.6.3",
-    "react-i18next": "^10.11.4",
-    "react-meta-tags": "^0.7.4",
-    "react-pose": "^4.0.6",
-    "react-router": "^5.0.0",
-    "react-router-dom": "^5.0.0",
-    "universal-cookie": "^3.1.0"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.1.6",
-    "@babel/plugin-proposal-class-properties": "^7.1.0",
-    "@babel/plugin-syntax-dynamic-import": "^7.0.0",
-    "@babel/preset-env": "^7.5.5",
-    "@babel/preset-react": "^7.0.0",
-    "babel-jest": "^24.8.0",
-    "babel-loader": "^8.0.4",
-    "clean-webpack-plugin": "^1.0.0",
-    "css-loader": "^1.0.1",
-    "extract-text-webpack-plugin": "^4.0.0-beta.0",
-    "file-loader": "^2.0.0",
-    "html-webpack-plugin": "^3.2.0",
-    "jest": "^24.8.0",
-    "node-sass": "^4.10.0",
-    "react-test-renderer": "^16.8.6",
-    "sass-loader": "^7.1.0",
-    "style-loader": "^0.23.1",
-    "webpack": "^4.25.1",
-    "webpack-cli": "^3.3.7",
-    "webpack-dev-server": "^3.2.1",
-    "webpack-merge": "^4.1.4"
-  }
-}
