+++ ui/src/static/css/fa_all.scss
@@ -7369,8 +7369,7 @@
     font-family: "Font Awesome 5 Brands";
     font-style: normal;
     font-weight: normal;
-    src: url(../font/fontawesome/fa-brands-400.eot);
-    src: url(../font/fontawesome/fa-brands-400.eot?#iefix) format("embedded-opentype"), url(../font/fontawesome/fa-brands-400.woff2) format("woff2"), url(../font/fontawesome/fa-brands-400.woff) format("woff"), url(../font/fontawesome/fa-brands-400.ttf) format("truetype"), url(../font/fontawesome/fa-brands-400.svg#fontawesome) format("svg")
+    src: url(../font/fontawesome/fa-brands-400.woff2) format("woff2"), url(../font/fontawesome/fa-brands-400.woff) format("woff"), url(../font/fontawesome/fa-brands-400.svg#fontawesome) format("svg")
 }
 
 .fab {
@@ -7381,8 +7380,7 @@
     font-family: "Font Awesome 5 Pro";
     font-style: normal;
     font-weight: 300;
-    src: url(../font/fontawesome/fa-light-300.eot);
-    src: url(../font/fontawesome/fa-light-300.eot?#iefix) format("embedded-opentype"), url(../font/fontawesome/fa-light-300.woff2) format("woff2"), url(../font/fontawesome/fa-light-300.woff) format("woff"), url(../font/fontawesome/fa-light-300.ttf) format("truetype"), url(../font/fontawesome/fa-light-300.svg#fontawesome) format("svg")
+    src: url(../font/fontawesome/fa-light-300.woff2) format("woff2"), url(../font/fontawesome/fa-light-300.woff) format("woff"), url(../font/fontawesome/fa-light-300.svg#fontawesome) format("svg")
 }
 
 .fal {
@@ -7393,8 +7391,7 @@
     font-family: "Font Awesome 5 Pro";
     font-style: normal;
     font-weight: 400;
-    src: url(../font/fontawesome/fa-regular-400.eot);
-    src: url(../font/fontawesome/fa-regular-400.eot?#iefix) format("embedded-opentype"), url(../font/fontawesome/fa-regular-400.woff2) format("woff2"), url(../font/fontawesome/fa-regular-400.woff) format("woff"), url(../font/fontawesome/fa-regular-400.ttf) format("truetype"), url(../font/fontawesome/fa-regular-400.svg#fontawesome) format("svg")
+    src: url(../font/fontawesome/fa-regular-400.woff2) format("woff2"), url(../font/fontawesome/fa-regular-400.woff) format("woff"), url(../font/fontawesome/fa-regular-400.svg#fontawesome) format("svg")
 }
 
 .fal,
@@ -7410,8 +7407,7 @@
     font-family: "Font Awesome 5 Pro";
     font-style: normal;
     font-weight: 900;
-    src: url(../font/fontawesome/fa-solid-900.eot);
-    src: url(../font/fontawesome/fa-solid-900.eot?#iefix) format("embedded-opentype"), url(../font/fontawesome/fa-solid-900.woff2) format("woff2"), url(../font/fontawesome/fa-solid-900.woff) format("woff"), url(../font/fontawesome/fa-solid-900.ttf) format("truetype"), url(../font/fontawesome/fa-solid-900.svg#fontawesome) format("svg")
+    src: url(../font/fontawesome/fa-solid-900.woff2) format("woff2"), url(../font/fontawesome/fa-solid-900.woff) format("woff"), url(../font/fontawesome/fa-solid-900.svg#fontawesome) format("svg")
 }
 
 .fa,
