--- ui/webpack.common.js.old
@@ -1,50 +1,0 @@
-"use strict";
-
-const path = require("path");
-
-require("@babel/polyfill");
-
-const CleanWebpackPlugin = require("clean-webpack-plugin");
-const HtmlWebpackPlugin = require("html-webpack-plugin");
-
-module.exports = {
-  entry: ["@babel/polyfill", "./src/app.js"],
-  output: {
-    filename: "[name].bundle.js",
-    path: path.resolve(__dirname, "dist"),
-    publicPath: "/"
-  },
-  module: {
-    rules: [
-      {
-        test: /\.(js|jsx)$/,
-        exclude: /node_modules/,
-        use: [
-          "babel-loader"
-        ]
-      },
-      {
-        test: /\.(png|svg|jpg|jpeg|gif)$/,
-        use: [
-          "file-loader"
-        ]
-      },
-      {
-        test: /\.(woff|woff2|eot|ttf|otf)$/,
-        use: [
-          "file-loader"
-        ]
-      }
-    ]
-  },
-  devServer: {
-    historyApiFallback: true
-  },
-  plugins: [
-    new CleanWebpackPlugin(["dist"]),
-    new HtmlWebpackPlugin({
-      title: "Feinwaru SPPD",
-      template: "./template.html"
-    })
-  ]
-};
